+++ blockchain/validate.go
@@ -1597,7 +1597,7 @@ func CheckTransactionInputs(subsidyCache *SubsidyCache, tx *dcrutil.Tx,
 		// Ensure the transaction is not spending coins which have not
 		// yet reached the required coinbase maturity.
 		coinbaseMaturity := int64(chainParams.CoinbaseMaturity)
-		originHeight := int64(utxoEntry.BlockHeight())
+		originHeight := utxoEntry.BlockHeight()
 		if utxoEntry.IsCoinBase() {
 			blocksSincePrev := txHeight - originHeight
 			if blocksSincePrev < coinbaseMaturity {
